Mongo CRUD applied and tested
Implemented CRUD with mongoDB. Made the Basic and Wind handlers into methods on a struct with config data, for easy access.
Split the extractData function into parseJSON and createJSON, as we need the parsed JSON for insertion into the db.
All the CRUD functions are methods on the config struct as well, so they can be called by the handlers, and also have access to the mongo Client